업데이트 버튼 미작동 -> 작동 앱 수준 빌드 그래들 설정 실수, API키 숨기기 정상 작동

+++ app/src/main/java/com/cc/air_quality/retrofit/AirQualityResponse.kt
@@ -30,7 +30,7 @@ data class AirQualityResponse(
                 val tp: Int,
                 val ts: String,
                 val wd: Int,
-                val ws: Int
+                val ws: Double
             )
         }
